Login bug fix
Issue: Login query was not being made.

Reason: In App.js localStorage token is being checked to route to Login or App. When updating the localStorage token after this (in Login.jsx) React does not assign the updated value in App.js as it is not assigned to a State type. So (!token) would return true in App.js routing the user back to the login page after login submission (before there is a chance to send the login mutation).

Fix: Restructured how the token is passed between components.